requirements.txt
wordcloud==1.8.1
python-bidi==0.4.2
arabic-reshaper==2.1.3
parsivar==0.2.3
